moved sensor_types and number_types to components
